TramosMapperPavimento uses CachedRowSet to insert tramos too.
The issue with this is the autoincrement/serial field (gid). As gid is NOT NULL,
CachedRowSet doesn't allow to create new records without explicity set a gid, so
I couldn't use typical strategies as write the INSERT without the gid or set gid
= NULL in order the database calculate it for me. Getting first the next value
of sequence and updating taking that into account works for me.