Fix bug on catalog: set the type of field caracteristica.
The tables in Catalog may contain different types of caracteristicas depending
on the layer (string if it's tipo_pavimento, double if it's ancho_plataforma,
etc). That is set when retrieving the records from DB taking into account that
info. But there was a problem when a road have no records (ie: a recently added
one), so we need to set the type in the TableModel.